refactor(core): Decouple post workflow execute event from internal hooks (no-changelog)
